requirements.txt
openai
lxml
pyyaml>=6.0
python-dotenv
